requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
httpx==0.27.2
python-dotenv==1.0.1
pydantic==2.10.4
tenacity==8.5.0
openai==1.59.7
